create letters.factories_test
